fix: https://github.com/open-runtime/runtime_keygen_openapi/issues/2
We need to be able to return different types from /me depending on the
type of token provided, but openapi-generator is just not capable of
generating the code to do that.

In theory, it should be possible to provide:
```
data:
  oneOf:
    $ref: "#/components/schemas/License"
    $ref: "#/components/schemas/Product"
    $ref: "#/components/schemas/User"
```
https://swagger.io/docs/specification/data-models/oneof-anyof-allof-not/

but the code that openapi-generator generates assumes a common type of
simply Object, but then does not update its JSON deserialization code
to understand Object, and the code throws different internal exceptions.

There are plenty of open anyOf-related bugs:
https://github.com/OpenAPITools/openapi-generator/issues?q=anyOf

So the strategy here is to:
1. remove the incorrect /me definition from keygen-openapi.yml
2. manually modify the generated code to return different types
3. update the .patch file